requirements.txt
Flask
yfinance
nselib
psycopg2-binary

